+++ src/Cattech.Optimizer.Pro.Infrastructure/Hardware/SensorProvider/IHardwareMonitorAbstraction.cs
@@ -22,6 +22,7 @@ internal enum InternalSensorType
     Temperature,
     Load,
     Clock,
+    SmallData,
     Other
 }